requirements.txt
Flask
torch
torchvision
torchmetrics
pillow
matplotlib
